- Fixed a possible timing bug in Linux and Mac - Fixed a possible crash bug - Fixed bug that made lightning effects not work - Fixed some texture filtering mode bug in setup

+++ game/sys_linux.c
@@ -15,7 +15,7 @@ double sys_getTime()
 {
   struct timeval now;
   gettimeofday( &now, NULL );
-  return now.tv_sec + now.tv_usec / 1000000.0f - startuptime;
+  return  ((double)now.tv_sec) + now.tv_usec / 1000000.0f - startuptime; 
 }
 
 int sys_frameStep()
